Donut King 3 Set Up
Fixed a bug in GameManager regarding spawning enemies after they died. Coroutine was causing the IEnumerator method to constantly look for index 0  in the deadTracker list (where the dead go) despite there being nothing. Added lines of code to fix that error.

Set up a scene with John's help to add music and have each button make noise when clicked on.

Made changes to the door script with John to make sure that the "Press [F] to Open the Door" text only appears while the door is closed.