Door Bug Fixed
Doors now hold a list and stays open if people are in the doorway and closes when people aren't. Doors now also check the deadTracker List when people die (called from GameManager's DeclareSelfDead method.) Anyone that dies tells the GameManager to let the doors know to check their list and to remove them from their specific doorway list if they were ever in a doorway.

This safety was added to avoid null references when checking lists in doors for Adding to list and Removing from list (both occurring during OnTriggerEnter and OnTriggerExit respectively.)

Also changed each door's box collider to not open automatically when a player spawns in that room. This prevents confusion with the door as Spawning does not necessarily count as OnTriggerEnter (OnTriggerEnter relies on the fact that an entity is making contact FROM outside the box, not suddenly appearing inside the box.)