New Scene and Shop System Integration Complete
Implemented and cleaned-up code to make sure the shop ui system works with the current scene in place. Also set all enemyAI scripts to null for the time being, as enemyAI is going to require a revamp anyway. Pushed PJ's NMODEL_Player into the prefab under the name NMODEL_Player2 in the small chance this may cause issues.