Fixed Bugs on Updating Player Money Count on start of Next Round, Donuts being dropped by airborne players in air, and made parts of the map walkable and unwalkable for enemies
Made sure to ensure that the player money count was being updated properly and added rule where each timeHeld with donut nets $100 (0 seconds nets just $100.) Also wrote code in GameManager where players that drop the donut in midair ensures donut is spawned on the ground below them, not in the air with them. Also made couch walkable so that if player dies while on the couch and drops the donut, enemies can still grab the donut from the couch. Made the receptionist desk completely unwalkable and unjumpable for enemies and players alike.