Fixed bug where donut wasn't spawning at the middle when new round starts and added smoke FX when entity dies
There was a bug where the donut wasn't spawning at the center of the map (where it's meant to be) at the start of the round and enemies weren't exactly staying still when spawning in a new round. Also created visual FX for the donut and The Donut King to make it easier for players to spot who has the donut. Also made map indicators bigger for players to actually use the map. Added smoke FX when players/enemies die.