Added MatchSettings Feedback
Updated instructions on the Create Match page to let players know the basics of joining and requirements to start a match. Added error messages that when NO players are present and user attempts to click on NEXT, a red error message prints "MINIMUM 1 REAL PLAYER." If one player is added, but no other players/bots are in the Player Slots and user attempts to hit Next, a red error message prints "NEED 1 BOT OR 2 REAL PLAYERS." An error sound also plays if user attempts to hit NEXT but requirements are not met.